REF: Rename LOCAL_METADATA_FILENAME to METADATA_FILENAME
