+++ octue/utils/metadata.py
@@ -8,7 +8,7 @@ from octue.utils.decoders import OctueJSONDecoder
 logger = logging.getLogger(__name__)
 
 
-LOCAL_METADATA_FILENAME = ".octue"
+METADATA_FILENAME = ".octue"
 
 
 def load_local_metadata_file(path):
@@ -25,7 +25,7 @@ def load_local_metadata_file(path):
             return json.load(f, cls=OctueJSONDecoder)
         except json.decoder.JSONDecodeError:
             logger.warning(
-                f"The {LOCAL_METADATA_FILENAME!r} metadata file at {path!r} is incorrectly formatted so no metadata "
+                f"The {METADATA_FILENAME!r} metadata file at {path!r} is incorrectly formatted so no metadata "
                 f"can be read from it. Please fix or delete it."
             )
             return {}
